radio_controller/tests/requirements.txt
mock==4.0.3
parameterized==0.8.1
pytest==6.2.3
requests-mock==1.8.0
coverage~=5.5
testing.postgresql==1.3.0
psycopg2-binary==2.8.6